requirements.txt
aiogram==3.26.0
asyncpg==0.31.0
pydantic==2.12.0
python-dotenv==1.1.1
